+++ plugins/azure-wakamiti-plugin/src/main/java/es/iti/wakamiti/azure/AzureReporter.java
@@ -4,7 +4,9 @@ import es.iti.commons.jext.Extension;
 import es.iti.wakamiti.api.extensions.Reporter;
 import es.iti.wakamiti.api.plan.NodeType;
 import es.iti.wakamiti.api.plan.PlanNodeSnapshot;
+import es.iti.wakamiti.api.util.Pair;
 import es.iti.wakamiti.api.util.WakamitiLogger;
+import es.iti.wakamiti.azure.internal.Util;
 import org.slf4j.Logger;
 
 import java.io.IOException;
@@ -16,6 +18,7 @@ import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.stream.Stream;
 
 
@@ -29,11 +32,15 @@ public class AzureReporter implements Reporter {
 
     private static final Logger LOGGER = WakamitiLogger.forClass(AzureReporter.class);
     public static final String AZURE_PLAN = "azurePlan";
+    public static final String AZURE_PLAN_ID = "azurePlanId";
     public static final String AZURE_SUITE = "azureSuite";
+    public static final String AZURE_SUITE_ID = "azureSuiteId";
     public static final String AZURE_TEST = "azureTest";
     public static final String AZURE_AREA = "azureArea";
     public static final String AZURE_ITERATION = "azureIteration";
     public static final String AZURE_TEST_ID = "azureTestId";
+
+
 
     private boolean disabled;
     private String host;
@@ -106,6 +113,7 @@ public class AzureReporter implements Reporter {
     @Override
     public void report(PlanNodeSnapshot result) {
 
+        System.out.println("azure report disabled:" +disabled);
         if (disabled) {
             return;
         }
@@ -120,41 +128,30 @@ public class AzureReporter implements Reporter {
         );
 
 
-        Map<AzurePlan,Map<AzureSuite,List<PlanNodeSnapshot>>> testCases = getTestCases(result, new HashMap<>());
+        Map<AzurePlan,Map<AzureSuite,List<PlanNodeSnapshot>>> testCases = getOrCreateTestCases(result, new HashMap<>(), api);
         if (testCases.isEmpty()) {
             return;
         }
 
+
         for (var testCaseEntry: testCases.entrySet()) {
 
             AzurePlan testPlan = testCaseEntry.getKey();
-            AzurePlan azurePlan = getAzurePlan(testPlan,api);
-            if (azurePlan == null) {
-                continue;
-            }
 
             Map<String,PlanNodeSnapshot> testPoints = new HashMap<>();
 
             for (var suiteEntry : testCaseEntry.getValue().entrySet()) {
-
-                String suiteName = suiteEntry.getKey().name();
-                AzureSuite azureSuite = getAzureSuite(azurePlan, suiteName, api);
-                if (azureSuite == null) {
-                    continue;
-                }
-
+                AzureSuite azureSuite = suiteEntry.getKey();
                 List<PlanNodeSnapshot> nodes = suiteEntry.getValue();
-
                 testPoints.putAll( getTestPointsStatus(azureSuite, nodes, api) );
-
             }
 
             String runID = api.createRun(
-                    azurePlan.id(),
-                    testPoints.keySet(),
-                    testPlan.name() + " - run by Wakamiti ",
-                    adjustTimeZone(result.getStartInstant()),
-                    adjustTimeZone(result.getFinishInstant())
+                testPlan.id(),
+                testPoints.keySet(),
+                testPlan.name() + " - run by Wakamiti ",
+                adjustTimeZone(result.getStartInstant()),
+                adjustTimeZone(result.getFinishInstant())
             );
             api.updateRunResults(runID,testPoints);
             attachFiles(runID, api);
@@ -188,28 +185,42 @@ public class AzureReporter implements Reporter {
 
 
 
-    private Map<AzurePlan,Map<AzureSuite,List<PlanNodeSnapshot>>> getTestCases(
+    private Map<AzurePlan,Map<AzureSuite,List<PlanNodeSnapshot>>> getOrCreateTestCases(
         PlanNodeSnapshot node,
-        Map<AzurePlan,Map<AzureSuite,List<PlanNodeSnapshot>>> result
+        Map<AzurePlan,Map<AzureSuite,List<PlanNodeSnapshot>>> result,
+        AzureApi api
     ) {
         boolean matchAzureTestCase = (testCasePerFeature ?
             node.getNodeType() == NodeType.AGGREGATOR && "feature".equals(node.getProperties().get("gherkinType")) :
             node.getNodeType() == NodeType.TEST_CASE
         );
+
         if (matchAzureTestCase && node.getTags().contains(azureTag)) {
-            String testPlan = property(node, AZURE_PLAN);
-            String suiteName = property(node, AZURE_SUITE);
-            String area = property(node, AZURE_AREA, null);
-            String iteration = property(node, AZURE_ITERATION, null);
-            if (testPlan != null && suiteName != null) {
-                AzurePlan azurePlan = new AzurePlan(testPlan,area,iteration);
+
+            Pair<String,String> testPlan = Util.getPropertyIdAndName(node, AZURE_PLAN);
+            List<Pair<String, String>> testSuitePath = Util.getListPropertyIdAndName(node, AZURE_SUITE);
+            String area = Util.property(node, AZURE_AREA, null);
+            String iteration = Util.property(node, AZURE_ITERATION, null);
+
+            if (testPlan != null && testSuitePath != null) {
+
+                AzurePlan azurePlan = getOrCreateAzurePlan(testPlan,area,iteration,api);
+                if (azurePlan == null) {
+                    return result;
+                }
+
+                AzureSuite azureSuite = getOrCreateAzureSuite(azurePlan, testSuitePath, api);
+                if (azureSuite == null) {
+                    return result;
+                }
+
                 Map<AzureSuite,List<PlanNodeSnapshot>> suites = result.computeIfAbsent(azurePlan, x -> new HashMap<>());
-                AzureSuite azureSuite = new AzureSuite(suiteName);
                 List<PlanNodeSnapshot> nodes = suites.computeIfAbsent(azureSuite, x -> new LinkedList<>());
                 nodes.add(node);
             }
+
         } else if (node.getChildren() != null){
-            node.getChildren().forEach(child -> getTestCases(child,result));
+            node.getChildren().forEach(child -> getOrCreateTestCases(child,result,api));
         }
         return result;
     }
@@ -218,27 +229,26 @@ public class AzureReporter implements Reporter {
 
 
 
+
+
+
     private Map<String,PlanNodeSnapshot> getTestPointsStatus(AzureSuite suite, List<PlanNodeSnapshot> planTestCases, AzureApi api) {
         Map<String,PlanNodeSnapshot> testPoints = new HashMap<>();
         for (PlanNodeSnapshot testCase : planTestCases) {
-            String testName = property(testCase, AZURE_TEST, testCase.getName());
-            String definedTestId = property(testCase, AZURE_TEST_ID, null);
-            String testCaseID;
-            if (definedTestId != null) {
-                if (!checkExistTestId(suite, definedTestId, testName, api)) {
-                    continue;
-                }
-                testCaseID = definedTestId;
-            } else {
-                testCaseID = getTestCase(suite, testName, api);
-            }
-            if (testCaseID == null) {
+
+            AzureTestCase azureTestCase = getOrCreateTestCase(
+                suite,
+                Util.getPropertyValueIdAndName(testCase, AZURE_TEST, null),
+                api
+            );
+
+            if (azureTestCase == null) {
                 continue;
             }
             testPoints.put(api.getTestPointID(
                 suite.plan().id(),
-                suite.id(),
-                testCaseID
+                suite.idPath(),
+                azureTestCase.id()
             ), testCase);
         }
         return testPoints;
@@ -246,13 +256,31 @@ public class AzureReporter implements Reporter {
 
 
 
-    private AzurePlan getAzurePlan(AzurePlan testPlan, AzureApi api) {
-        return api.getPlan(testPlan.name(), testPlan.area(), testPlan.iteration()).orElseGet(()-> {
+
+
+    private AzurePlan getOrCreateAzurePlan(Pair<String,String> nameAndId, String area, String iteration, AzureApi api) {
+
+        Optional<AzurePlan> azurePlan;
+
+        String planId = nameAndId.value();
+        String planName = nameAndId.key();
+
+        if (planId != null) {
+            azurePlan = api.getPlanById(planId);
+            azurePlan.ifPresentOrElse(
+                it -> api.updatePlanName(planId, planName),
+                ()-> LOGGER.warn("Test Plan id {} not present in Azure", planId)
+            );
+        } else {
+            azurePlan = api.getPlanByProperties(planName, area, iteration);
+        }
+
+        return azurePlan.orElseGet(()-> {
             if (createItemsIfAbsent) {
-                LOGGER.info("Creating new test plan '{}' [ {} / {} ]", testPlan.name(), testPlan.area(), testPlan.iteration());
-                return api.createPlan(testPlan.name(), testPlan.area(), testPlan.iteration());
+                LOGGER.info("Creating new test plan '{}' [ {} / {} ]", planName, area, iteration);
+                return api.createPlan(planName, area, iteration);
             } else {
-                LOGGER.warn("Test plan '{}' [ {} / {} ] is not defined and will be ignored", testPlan.name(), testPlan.area(), testPlan.iteration());
+                LOGGER.warn("Test plan '{}' [ {} / {} ] is not defined and will be ignored", planName, area, iteration);
                 return null;
             }
         });
@@ -260,63 +288,87 @@ public class AzureReporter implements Reporter {
 
 
 
-    private AzureSuite getAzureSuite(AzurePlan azurePlan, String suiteName, AzureApi api) {
-        return api.getTestSuite(azurePlan, suiteName).orElseGet(()-> {
-            if (createItemsIfAbsent) {
-                LOGGER.info("Creating new test suite '{}'", suiteName);
-                return api.createSuite(azurePlan, suiteName);
+    private AzureSuite getOrCreateAzureSuite(AzurePlan azurePlan, List<Pair<String,String>> suitePath, AzureApi api) {
+
+        AzureSuite parent = null;
+        AzureSuite azureSuite = null;
+
+        for (Pair<String,String> nameAndId : suitePath) {
+
+            LOGGER.debug("getOrCreateAzureSuite (path = [{}] {})", nameAndId.value(), nameAndId.key());
+
+            String suiteName = nameAndId.key();
+            String suiteId = nameAndId.value();
+
+            if (suiteId != null) {
+                azureSuite = api.getTestSuiteById(azurePlan,suiteId, parent).orElse(null);
+                if (azureSuite != null) {
+                    api.updateTestSuiteName(azurePlan, suiteId, suiteName);
+                } else {
+                    LOGGER.warn("Test Suite id {} not present in Azure", suiteId);
+                }
             } else {
-                LOGGER.warn("Test suite '{}' is not defined and will be ignored", suiteName);
-                return null;
+                azureSuite = api.getTestSuiteByName(azurePlan, suiteName, parent).orElse(null);
             }
-        });
+
+            if (azureSuite == null) {
+                if (createItemsIfAbsent) {
+                    LOGGER.info("Creating new test suite '{}'", suiteName);
+                    azureSuite = api.createSuite(azurePlan, suiteName, parent);
+                } else {
+                    LOGGER.warn("Test suite '{}' is not defined and will be ignored", suiteName);
+                    return null;
+                }
+            }
+
+            parent = azureSuite;
+        }
+
+        return azureSuite;
     }
 
 
 
-    private String getTestCase(AzureSuite suite, String testName, AzureApi api) {
-        return api.getTestCaseID(suite,testName).orElseGet(()-> {
-            if (createItemsIfAbsent) {
-                LOGGER.info("Creating new work item for test case '{}'", testName);
-                return api.createTestCase(suite,testName);
+
+
+
+    private AzureTestCase getOrCreateTestCase(AzureSuite suite, Pair<String,String> nameAndId, AzureApi api) {
+
+        String testCaseId = nameAndId.value();
+        String testCaseName = nameAndId.key();
+        AzureTestCase azureTestCase;
+
+        if (testCaseId != null) {
+            if (!api.existsTestCaseID(suite,testCaseId)) {
+                LOGGER.warn("Test Case id '{}' not present in Azure", testCaseId);
+                return null;
             } else {
-                LOGGER.warn("Test case '{}' is not defined in Azure test plan and will be ignored", testName);
-                return null;
+                api.updateTestCaseName(testCaseId, testCaseName);
+                azureTestCase = new AzureTestCase(testCaseId, testCaseName);
             }
-        });
-    }
-
-
-
-
-    private boolean checkExistTestId(AzureSuite suite, String definedTestId, String testName, AzureApi api) {
-        try {
-            api.getTestPointID(suite.plan().id(), suite.id(), definedTestId);
-            api.updateTestCaseName(definedTestId, testName);
-            return true;
-        } catch (Exception e) {
-            return false;
+        } else {
+            azureTestCase = api.getTestCaseByName(suite, testCaseName).orElse(null);
         }
-    }
 
-
-
-    private String property(PlanNodeSnapshot node, String property) {
-        if (node.getProperties() == null || !node.getProperties().containsKey(property)) {
-            LOGGER.warn("Property {} not present in test case {}", property, node.getDisplayName());
-            return null;
+        if (testCaseId == null) {
+            if (createItemsIfAbsent) {
+                LOGGER.info("Creating new test case '{}'", testCaseName);
+                azureTestCase = api.createTestCase(suite, testCaseName);
+            } else {
+                LOGGER.warn("Test case '{}' is not defined and will be ignored", testCaseName);
+            }
         }
-        return node.getProperties().get(property);
+
+        return azureTestCase;
+
     }
 
 
 
-    private String property(PlanNodeSnapshot node, String property, String defaultValue) {
-        if (node.getProperties() == null || !node.getProperties().containsKey(property)) {
-            return defaultValue;
-        }
-        return node.getProperties().get(property);
-    }
+
+
+
+
 
 
 }