+++ plugins/file-uploader-wakamiti-plugin/src/main/java/es/iti/wakamiti/fileuploader/FilesUploaderConfigurator.java
@@ -22,6 +22,7 @@ public class FilesUploaderConfigurator implements ConfigContributor<AbstractFile
     private static final String PASSWORD = "credentials.password";
     private static final String DESTINATION_DIR = "destinationDir";
     private static final String PROTOCOL = "protocol";
+    private static final String IDENTITY = "identity";
 
 
     @Override
@@ -55,6 +56,7 @@ public class FilesUploaderConfigurator implements ConfigContributor<AbstractFile
             conf.get(PASSWORD, String.class).ifPresent(filesUploader::setPassword);
             conf.get(DESTINATION_DIR, String.class).ifPresent(filesUploader::setRemotePath);
             conf.get(PROTOCOL, String.class).ifPresent(filesUploader::setProtocol);
+            conf.get(IDENTITY, String.class).ifPresent(filesUploader::setIdentity);
         }
     }