+++ plugins/file-uploader-wakamiti-plugin/CHANGELOG.md
@@ -12,6 +12,15 @@ and this project adheres to [Semantic Versioning][2].
 - Upgrade dependencies: `wakamiti-plugin-starter`
 
 
+## [2.5.0] - 2023-11-22
+
+### Added
+- Support for using `ppk` identity files
+
+### Changed
+- Replace dependency `com.jcraft:jsch:0.1.55` with `com.github.mwiede:jsch:0.2.13`
+
+
 ## [2.4.3] - 2023-09-14
 
 ### Fixed